Fix bug where logout => login causes tabs to break
