+++ src/TabbedRoutes.tsx
@@ -24,11 +24,10 @@ import { jwtIssSelector, jwtSelector } from "./features/auth/authSlice";
 import ActorRedirect from "./ActorRedirect";
 import SpecialFeedPage from "./pages/shared/SpecialFeedPage";
 import styled from "@emotion/styled";
-import ProfilePage from "./pages/profile/ProfilePage";
+import UserPage from "./pages/profile/UserPage";
 import SettingsPage from "./pages/settings/SettingsPage";
 import { useContext, useRef } from "react";
 import { AppContext } from "./features/auth/AppContext";
-import UserPage from "./pages/shared/UserPage";
 import InstallAppPage from "./pages/settings/InstallAppPage";
 import SearchPage from "./pages/search/SearchPage";
 import SearchPostsResultsPage from "./pages/search/results/SearchFeedResultsPage";
@@ -53,6 +52,7 @@ import AppearancePage from "./pages/settings/AppearancePage";
 import CommunitySidebarPage from "./pages/shared/CommunitySidebarPage";
 import ApolloMigratePage from "./pages/settings/ApolloMigratePage";
 import PostAppearancePage from "./pages/settings/PostAppearancePage";
+import ProfilePage from "./pages/profile/ProfilePage";
 
 const Interceptor = styled.div`
   position: absolute;
@@ -288,7 +288,7 @@ export default function TabbedRoutes() {
           {...buildGeneralBrowseRoutes("inbox")}
 
           <Route exact path="/profile">
-            <ProfilePage key={jwt} />
+            <ProfilePage />
           </Route>
           {...buildGeneralBrowseRoutes("profile")}
